soo many changes
First off, i deleted all the images and added them to spriteSheets, now there is a spriteReader objet that given a spritesheet and info can return the correct sprites from that sheet. There are now also 3 clases that handles the priteLoading. That is the BDMap and also the clases BDBug and BDPlayer, that is because they have a lot of sprites themselves. Might change this later.

I also realized that the poor bastards that are going to read all my files and give me a grade might not like the fact that i changed the name of BDBug to BDMonster, so i changed it back.

Lastely I added a door so that there is a goal with the game. This is because I wish to add an AI later and it will need a goal. On entering the door there will be given a score depending on the amount of diamonds you got, monstrers not killed and the time left. This also added a timer that is counting down, although the timer is not super accurate and depend on ingame time not system time it does still work.

Yeah, I think that was it :P